backend/agent/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aniso8601==10.0.0
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
copilotkit==0.1.41
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.6
finnhub-python==2.4.23
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
flask-restx==1.3.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.14
langchain-anthropic==0.2.4
langchain-community==0.3.14
langchain-core==0.3.30
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.64
langgraph-api==0.0.32
langgraph-checkpoint==2.0.21
langgraph-cli==0.1.79
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.2.11
lxml==4.9.4
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.59.8
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
partialjson==0.0.8
peewee==3.17.9
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
rpds-py==0.23.1
# Editable install with no version control (sample_agent==0.0.1)
-e /Users/bryantan/Documents/FYP Final/HengOngHuat/agent_test/CopilotKit/examples/coagents-starter/agent-py
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.3
structlog==25.2.0
tenacity==8.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.31.1
watchfiles==1.0.4
wcwidth==0.2.13
Werkzeug==3.1.3
yahooquery==2.3.7
yarl==1.18.3
yfinance==0.2.55
zstandard==0.23.0
